feat(control_editor): increase position and size precision
This commit increases the precision of widget position and size values
by a factor of 10. This change allows for finer control over widget
placement and dimensions.

The editor version has been bumped to 2 to reflect this change.
Layout files saved with editor version 1 will be automatically
updated to the new precision when loaded.